# Reuse context in compact unwind table generation